% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geoda.R
\name{as.data.frame.geoda}
\alias{as.data.frame.geoda}
\title{convert rgeoda instance to data.frame}
\usage{
\method{as.data.frame}{geoda}(gda_obj)
}
\arguments{
\item{gda_obj}{A rgeoda object}
}
\value{
A data.frame object
}
\description{
Override the as.data.frame function for rgeoda instance
}
