% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_geoda.R
\name{as.geoda}
\alias{as.geoda}
\title{Create an instance of geoda-class from either an 'sf' or 'sp' object}
\usage{
as.geoda(obj)
}
\arguments{
\item{obj}{An instance of 'sf' or 'sp' object}
}
\value{
An instance of geoda-class
}
\description{
Create an instance of geoda-class from an 'sf' object returned from 'st_read()' function,
or a 'sp' object returned from 'readOGR()' function.
NOTE: The table content is NOT used to create an instance of geoda-class.
}
