% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{as.matrix}
\alias{as.matrix}
\title{as.matrix}
\usage{
as.matrix(obj)
}
\arguments{
\item{obj}{A weights object}
}
\value{
A matrix object
}
\description{
Convert a spatial weights object to a Matrix object
}
