% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{between_sumofsquare}
\alias{between_sumofsquare}
\title{Between Sum of Square}
\usage{
between_sumofsquare(clusters, data)
}
\arguments{
\item{clusters}{A 2d list which returns from spatial clustering methods, e.g. skater()}

\item{data}{A 2d list which is used in spatial clustering methods, e.g. skater(k, w, data)}
}
\value{
A value of between sum of square value
}
\description{
Compute between sum of square value of a group of clusters
}
