% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoda.R
\name{$,p_GeoDa-method}
\alias{$,p_GeoDa-method}
\title{$, expose member functions of p_GeoDa class
e.g. guerry$GetNumObs()}
\usage{
\S4method{$}{p_GeoDa}(x, name)
}
\description{
$, expose member functions of p_GeoDa class
e.g. guerry$GetNumObs()
}
