% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoda.R
\name{$,p_GeoDaWeight-method}
\alias{$,p_GeoDaWeight-method}
\title{$, expose member functions of p_GeoDaWeight class}
\usage{
\S4method{$}{p_GeoDaWeight}(x, name)
}
\description{
$, expose member functions of p_GeoDaWeight class
}
