% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{create_weights}
\alias{create_weights}
\title{Create an empty weights}
\usage{
create_weights(num_obs)
}
\arguments{
\item{num_obs}{The number of observations for this empty weights}
}
\value{
An instance of Weight-class
}
\description{
Create an empty weights
}
