% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eb_rates}
\alias{eb_rates}
\title{Empirical Bayes(EB) Rate}
\usage{
eb_rates(event_data, base_data)
}
\arguments{
\item{event_data}{A numeric array of selected "event" variable}

\item{base_data}{A numeric array of selected "base" variable}
}
\value{
A data.frame with two columns "EB Rate" and "IsNull".
}
\description{
The function to compute EB Rate from an event variable and a base variable.
}
\examples{
\dontrun{
nat <- geoda_open("natregimes.shp")
nat_df <- as.data.frame(nat)
hr60 <- guerry_df['HR60'][,1]
po60 <- guerry_df['PO60'][,1]
ebr <- eb_rates(hr60, po60)
ebr
}
}
