% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{flat_clusters}
\alias{flat_clusters}
\title{Flat a group of clusters to an integer array}
\usage{
flat_clusters(clusters)
}
\arguments{
\item{clusters}{A list of numeric vectors which returns from spatial clustering methods, e.g. skater()}
}
\value{
An integer array
}
\description{
Flat the result of a spatial clustering method (e.g. skater, redcap etc.) to an integer array
}
