% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{gda_min_distthreshold}
\alias{gda_min_distthreshold}
\title{(For internally use and test only) Minimum Distance Threshold for Distance-based Weights}
\usage{
gda_min_distthreshold(geoda_obj, is_arc = FALSE, is_mile = TRUE)
}
\arguments{
\item{geoda_obj}{An instance of geoda-class}

\item{is_arc}{(optional) FALSE (default) or TRUE, compute arc distance between two observations}

\item{is_mile}{(optional) TRUE (default) or FALSE, if 'is_arc' option is TRUE, then 'is_mile' will set distance unit to 'mile' or 'km'.}
}
\value{
A numeric value of minimum threshold of distance
}
\description{
Get minimum threshold of distance that makes sure each observation has at least one neighbor
}
