% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geoda.R
\docType{class}
\name{geoda-class}
\alias{geoda-class}
\alias{geoda}
\title{A R wrapper for GeoDa}
\arguments{
\item{o_gda}{An object of _p_GeoDa class}
}
\value{
An object of geoda class
}
\description{
geoda is a RefClass that wraps the C++ GeoDa class (via _p_GeoDa defines in rgeoda.R)
}
\section{Methods}{

\describe{
\item{\code{GetIntegerCol(col_name)}}{Get the values (in integer type) from a column}

\item{\code{GetRealCol(col_name)}}{Get the values (in real type) from a column}
}}

