% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geoda.R
\name{geoda_open}
\alias{geoda_open}
\title{Create an instance of geoda-class by reading from an ESRI Shapefile dataset}
\usage{
geoda_open(ds_path)
}
\arguments{
\item{ds_path}{(character) The path of the spatial dataset}
}
\value{
An instance of geoda-class
}
\description{
Create an instance of geoda-class by reading from an ESRI Shapefile dataset.
}
\examples{
guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- geoda_open(guerry_path)
guerry_df <- as.data.frame(guerry) # access as a data.frame
head(guerry_df)
}
