% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_kcomponents}
\alias{get_kcomponents}
\title{Get first k components from PCA}
\usage{
get_kcomponents(pca_r, k)
}
\arguments{
\item{pca_r}{An object return from pca() function}

\item{k}{A numeric value}
}
\value{
A PCAResult object containing the details of PCA
}
\description{
Get first k components from PCA
}
