% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoda.R
\name{initialize,p_GeoDaWeight-method}
\alias{initialize,p_GeoDaWeight-method}
\title{Constructors for p_GeoDaWeight class
Note: here simply using argc to determine which constructor should be called}
\usage{
\S4method{initialize}{p_GeoDaWeight}(.Object, ...)
}
\description{
Constructors for p_GeoDaWeight class
Note: here simply using argc to determine which constructor should be called
}
