% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{is_symmetric}
\alias{is_symmetric}
\title{Symmetry of Weights Matrix}
\usage{
is_symmetric(gda_w)
}
\arguments{
\item{gda_w}{A Weight object}
}
\value{
A boolean value indicates if weights matrix is symmetric
}
\description{
Check if weights matrix is symmetric
}
\examples{
\dontrun{
guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- st_read(guerry_path)
queen_w <- queen_weights(guerry)
is_symmetric(queen_w)
}
}
