% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{lisa_bo}
\alias{lisa_bo}
\title{Get Bonferroni bound value of local spatial autocorrelation}
\usage{
lisa_bo(gda_lisa, current_p)
}
\arguments{
\item{gda_lisa}{An instance of LISA object}

\item{current_p}{A value of current siginificant p-value}
}
\value{
A value of Bonferroni bound
}
\description{
Get Bonferroni bound value based on current LISA computation and current significat p-value
}
