% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{lisa_colors}
\alias{lisa_colors}
\title{Get cluster colors}
\usage{
lisa_colors(gda_lisa)
}
\arguments{
\item{gda_lisa}{An instance of LISA object}
}
\value{
colors a tuple of values
}
\description{
Get the cluster colors of LISA computation.
}
