% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{lisa_pvalues}
\alias{lisa_pvalues}
\title{Get values of local pseudo-p values}
\usage{
lisa_pvalues(gda_lisa)
}
\arguments{
\item{gda_lisa}{An instance of LISA object}
}
\value{
data a tuple of pseudo-p values of local spatial autocorrelation
}
\description{
Get the local pseudo-p values of significance returned from LISA computation.
}
