% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{lisa_values}
\alias{lisa_values}
\title{Get values of local spatial autocorrelation}
\usage{
lisa_values(gda_lisa)
}
\arguments{
\item{gda_lisa}{An instance of LISA object}
}
\value{
data a tuple of values of local spatial autocorrelation
}
\description{
Get the local spatial autocorrelation values returned from LISA computation
}
