% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_g}
\alias{local_g}
\title{Local Getis-Ord's G statistics}
\usage{
local_g(w, data)
}
\arguments{
\item{w}{An instance of Weight object}

\item{data}{A tuple of values of selected variable}
}
\value{
lisa_obj An instance of LISA (LocalSpatialAutocorrelation) object
}
\description{
The function to apply Getis-Ord's local G statistics
}
