% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_geary}
\alias{local_geary}
\title{Local Geary statistics}
\usage{
local_geary(w, data)
}
\arguments{
\item{w}{An instance of Weight object}

\item{data}{A tuple of values of selected variable}
}
\value{
lisa_obj An instance of LISA (LocalSpatialAutocorrelation) object
}
\description{
The function to apply local Geary statistics
}
