% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_gstar}
\alias{local_gstar}
\title{Local Getis-Ord's G* statistics}
\usage{
local_gstar(w, data, ncpu = 8, perm = 999)
}
\arguments{
\item{w}{An instance of Weight object}

\item{data}{A tuple of values of selected variable}

\item{ncpu}{The number of how many cpu threads being used, default: 8}

\item{perm}{The number of permutations used for permutation inference, default: 999}
}
\value{
lisa_obj An instance of LISA (LocalSpatialAutocorrelation) object
}
\description{
The function to apply Getis-Ord's local G* statistics
}
