% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_joincount}
\alias{local_joincount}
\title{Local Join Count statistics}
\usage{
local_joincount(w, data, ncpu = 8, perm = 999)
}
\arguments{
\item{w}{An instance of Weight object}

\item{data}{A tuple of values of selected variable}
}
\value{
lisa_obj An instance of LISA (LocalSpatialAutocorrelation) object
}
\description{
The function to apply local Join Count statistics
}
