% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losh.R
\name{local_losh}
\alias{local_losh}
\title{Local Spatial Heteroscedasticity (LOSH)}
\usage{
local_losh(
  w,
  df,
  a = 2,
  permutations = 999,
  permutation_method = "complete",
  significance_cutoff = 0.05,
  cpu_threads = 6,
  seed = 123456789
)
}
\arguments{
\item{w}{An instance of Weight object}

\item{df}{A data frame with only selected variable. E.g. guerry["Crm_prs"]}

\item{a}{(optional) The power (or exponent) of the absolute difference.
Default is 2.}

\item{permutations}{(optional) The number of permutations for the LISA
computation}

\item{permutation_method}{(optional) The permutation method used for the
LISA computation. Options are ('complete', 'lookup'). Default is 'complete'.}

\item{significance_cutoff}{(optional) A cutoff value for significance
p-values to filter not-significant clusters}

\item{cpu_threads}{(optional) The number of cpu threads used for parallel
LISA computation}

\item{seed}{(optional) The seed for random number generator}
}
\value{
An instance of LISA-class
}
\description{
The function to apply local spatial heteroscedasticity (LOSH)
statistics. It identifies areas of local spatial heteroscedasticity,
which can be used to characterize the internal heterogeneity of hotspots.
The cluster indicators are: 0 (Not significant), 1 (Heterogeneous), 
and 2 (Homogeneous). Indices 5 (Undefined) and 6 (Isolated) are also used.
}
\examples{
\dontrun{
library(sf)
guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- st_read(guerry_path)
queen_w <- queen_weights(guerry)
losh <- local_losh(queen_w, guerry["Crm_prs"])
Hi <- lisa_values(losh)
}
}
