% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_moran_eb}
\alias{local_moran_eb}
\title{Local Moran with Empirical Bayes(EB) Rate}
\usage{
local_moran_eb(
  w,
  event_data,
  base_data,
  permutations = 999,
  significance_cutoff = 0.05,
  cpu_threads = 6,
  seed = 123456789
)
}
\arguments{
\item{w}{An instance of Weight object}

\item{event_data}{A numeric array of selected "event" variable}

\item{base_data}{A numeric array of selected "base" variable}

\item{permutations}{The number of permutations for the LISA computation}

\item{significance_cutoff}{A cutoff value for significance p-values to filter not-significant clusters}

\item{cpu_threads}{The number of cpu threads used for parallel LISA computation}

\item{seed}{The seed for random number generator}
}
\value{
An instance of LISA-class
}
\description{
The function to apply local Moran with EB Rate statistics. The EB rate is first computed from "event" and "base" variables, and then used in local moran statistics.
}
\examples{
\dontrun{
nat <- geoda_open("natregimes.shp")
nat_w <- queen_weights(nat)
nat_df <- as.data.frame(nat)
hr60 <- guerry_df['HR60'][,1]
po60 <- guerry_df['PO60'][,1]
lisa <- local_moran_eb(queen_w, hr60, po60)
lms <- lisa_values(lisa)
lms
}
}
