% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_multigeary}
\alias{local_multigeary}
\title{Local Multivariate Geary statistics}
\usage{
local_multigeary(w, data, ncpu = 8, perm = 999)
}
\arguments{
\item{w}{An instance of Weight object}

\item{data}{A 2D tuple of values of selected variables}
}
\value{
lisa_obj An instance of LISA (LocalSpatialAutocorrelation) object
}
\description{
The function to apply local Multivariate Geary statistics
}
