% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_multijoincount}
\alias{local_multijoincount}
\title{Local Multivariate Join Count statistics}
\usage{
local_multijoincount(w, data, ncpu = 8, perm = 999)
}
\arguments{
\item{w}{An instance of Weight object}

\item{data}{A data.frame with selected variables}
}
\value{
lisa_obj An instance of LISA (LocalSpatialAutocorrelation) object
}
\description{
The function to apply local Multivariate Join Count statistics
}
