% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_multiquantilelisa}
\alias{local_multiquantilelisa}
\title{Multivariate Quantile LISA Statistics}
\usage{
local_multiquantilelisa(
  w,
  df,
  ks,
  qs,
  permutations = 999,
  significance_cutoff = 0.05,
  cpu_threads = 6,
  seed = 123456789
)
}
\arguments{
\item{w}{An instance of Weight object}

\item{df}{A data frame with selected variables only. E.g. guerry[c("TopCrm", "TopWealth", "TopLit")]}

\item{ks}{A vector of "k" values indicate the number of quantiles for each variable. Value range e.g. [1, 10]}

\item{qs}{A vector of "q" values indicate which quantile or interval for each variable used in local join count statistics. Value stars from 1.}

\item{permutations}{The number of permutations for the LISA computation}

\item{significance_cutoff}{A cutoff value for significance p-values to filter not-significant clusters}

\item{cpu_threads}{The number of cpu threads used for parallel LISA computation}

\item{seed}{The seed for random number generator}
}
\value{
An instance of LISA-class
}
\description{
The function to apply multivariate quantile LISA statistics
}
