% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_quantilelisa}
\alias{local_quantilelisa}
\title{Quantile LISA statistics}
\usage{
local_quantilelisa(w, k, q, data, ncpu = 8, perm = 999)
}
\arguments{
\item{w}{An instance of Weight object}

\item{k}{A value indicates the number of quantiles. Value range e.g. [1, 10]}

\item{q}{A value indicates which quantile or interval used in local join count statistics. Value stars from 1.}

\item{data}{A tuple of values of selected variable}
}
\value{
lisa_obj An instance of LISA (LocalSpatialAutocorrelation) object
}
\description{
The function to apply quantile LISA statistics
}
