% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{max_neighbors}
\alias{max_neighbors}
\title{Maximum Neighbors of Spatial Weights}
\usage{
max_neighbors(gda_w)
}
\arguments{
\item{gda_w}{A Weight object}
}
\value{
The number of maximum neighbors of spatial weights
}
\description{
Get the number of maximum neighbors of spatial weights
}
\examples{
\dontrun{
guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- geoda_open(guerry_path)
queen_w <- queen_weights(guerry)
max_neighbors(queen_w)
}
}
