% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{mds}
\alias{mds}
\title{Multi Dimensional Scaling}
\usage{
mds(data, k)
}
\arguments{
\item{data}{A 2d numeric list of selected variables}

\item{k}{The number of clusters}
}
\value{
A 2d list represents a group of clusters
}
\description{
Apply multi dimensional scaling to project original data on to a low dimension (k) space
}
