% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{min_distthreshold}
\alias{min_distthreshold}
\title{Get minimum threshold of distance that makes sure each observation has at least one neighbor}
\usage{
min_distthreshold(geoda_obj, ...)
}
\arguments{
\item{is_arc}{(optional) FALSE (default) or TRUE, compute arc distance between two observations}

\item{is_mile}{(optional) TRUE (default) or FALSE, convert distance unit from mile to km.}
}
\value{
dist A real value of minimum threshold of distance
}
\description{
Get minimum threshold of distance that makes sure each observation has at least one neighbor
}
