% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{min_neighbors}
\alias{min_neighbors}
\title{Minimum Neighbors of Spatial Weights}
\usage{
min_neighbors(gda_w)
}
\arguments{
\item{gda_w}{A Weight object}
}
\value{
The number of minimum neighbors of spatial weights
}
\description{
Get the number of minimum neighbors of spatial weights
}
\examples{
\dontrun{
guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- geoda_open(guerry_path)
queen_w <- queen_weights(guerry)
min_neighbors(queen_w)
}
}
