% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{neighbor_match_test}
\alias{neighbor_match_test}
\title{Local Neighbor Match Test}
\usage{
neighbor_match_test(
  geoda_obj,
  k,
  data,
  scale_method = "standardize",
  distance_type = "euclidean",
  power = 1,
  is_inverse = FALSE,
  is_arc = FALSE,
  is_mile = TRUE
)
}
\arguments{
\item{geoda_obj}{An instance of geoda.}

\item{k}{a positive integer number for k-nearest neighbors searching.}

\item{data}{A list of numeric values of selected variables.}

\item{scale_method}{One of the scaling methods {'standardize', 'demean', 'mad', 'range_standardize', 'range_adjust'} to apply on input data. Default is 'standardize' (Z-score normalization).}

\item{distance_type}{The type of distance metrics used to measure the distance between input data. Options are {'euclidean', 'manhattan'}. Default is 'euclidean'.}

\item{power}{(optional) The power (or exponent) of a number says how many times to use the number in a multiplication.}

\item{is_inverse}{(optional) FALSE (default) or TRUE, apply inverse on distance value.}

\item{is_arc}{(optional) FALSE (default) or TRUE, compute arc distance between two observations.}

\item{is_mile}{(optional) TRUE (default) or FALSE, convert distance unit from mile to km.}
}
\value{
A data.frame with two columns "Cardinality" and "Probability".
}
\description{
The local neighbor match test is to assess the extent of overlap between k-nearest neighbors in geographical space and k-nearest neighbors in multi-attribute space.
}
\examples{
guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- geoda_open(guerry_path)
guerry_df <- as.data.frame(guerry) # use as data.frame
data <- guerry_df[c('Crm_prs','Crm_prp','Litercy','Donatns','Infants','Suicids')]
nbr_test <- neighbor_match_test(guerry, 6, data)
nbr_test
}
