% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_gda_demean}
\alias{p_gda_demean}
\title{p_gda_demean}
\usage{
p_gda_demean(data)
}
\arguments{
\item{data}{An input data for median absolute deviation}
}
\value{
A list of numeric vectors
}
\description{
The mean for each variable is subtracting from each observation resulting in mean zero.
}
