% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_gda_distance_weights}
\alias{p_gda_distance_weights}
\title{p_gda_distance_weights}
\usage{
p_gda_distance_weights(
  xp_geoda,
  dist_thres,
  power,
  is_inverse,
  is_arc,
  is_mile
)
}
\value{
An object of GeoDaWeight class
}
\description{
p_gda_distance_weights
}
