% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_gda_kernel_knn_weights}
\alias{p_gda_kernel_knn_weights}
\title{p_kernel_knn_weights}
\usage{
p_gda_kernel_knn_weights(
  xp_geoda,
  k,
  power,
  is_inverse,
  is_arc,
  is_mile,
  kernel_method,
  bandwidth,
  adaptive_bandwidth,
  use_kernel_diagonals
)
}
\value{
An object of GeoDaWeight class
}
\description{
p_kernel_knn_weights
}
