% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_gda_min_distthreshold}
\alias{p_gda_min_distthreshold}
\title{p_gda_min_distthreshold}
\usage{
p_gda_min_distthreshold(xp_geoda, is_arc, is_mile)
}
\value{
A real value of minimum threshold of distance
}
\description{
Return Get minimum threshold of distance
}
