% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_gda_queen_weights}
\alias{p_gda_queen_weights}
\title{p_gda_queen_weights}
\usage{
p_gda_queen_weights(xp_geoda, order, include_lower_order, precision_threshold)
}
\value{
An object of GeoDaWeight class
}
\description{
p_gda_queen_weights
}
