% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_gda_standardize}
\alias{p_gda_standardize}
\title{p_gda_standardize}
\usage{
p_gda_standardize(data)
}
\arguments{
\item{data}{An input data for median absolute deviation}
}
\value{
A list of numeric vectors
}
\description{
Standarize data by transforming data to have zero mean and unit variance
}
