% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_gda_standardize_mad}
\alias{p_gda_standardize_mad}
\title{p_gda_standardize_mad}
\usage{
p_gda_standardize_mad(data)
}
\arguments{
\item{data}{An input data for median absolute deviation}
}
\value{
A list of numeric vectors
}
\description{
Median absolute deviation to measure  measure of the variability of a univariate sample of quantitative data
}
