% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoda.R
\docType{class}
\name{p_GeoDa-class}
\alias{p_GeoDa-class}
\alias{p_GeoDa}
\title{p_GeoDa}
\description{
p_GeoDa class is a RefClass that wraps the C++ GeoDa class
See C++ functions in rcpp_rgeoda.cpp
}
\details{
This file is used to wrap C++ classes and functions defines in RcppExports.R
All other R script files will use this file as a bridge to C++ classes and functions

Author: lixun910@gmail.com
Changes:
10/29/2020 Add p_GeoDa class
}
