% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDa__GetFieldNames}
\alias{p_GeoDa__GetFieldNames}
\title{p_GeoDa__GetFieldNames}
\usage{
p_GeoDa__GetFieldNames(xp)
}
\value{
A list of field names
}
\description{
Get the field names from GeoDa object
}
