% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDa__GetFieldTypes}
\alias{p_GeoDa__GetFieldTypes}
\title{p_GeoDa__GetFieldTypes}
\usage{
p_GeoDa__GetFieldTypes(xp)
}
\value{
A list of field types
}
\description{
Get the field types from GeoDa object
}
