% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDa__GetIntegerCol}
\alias{p_GeoDa__GetIntegerCol}
\title{p_GeoDa__GetIntegerCol}
\usage{
p_GeoDa__GetIntegerCol(xp, col_name)
}
\value{
A list of integer values
}
\description{
Get the values of a integer col from GeoDa object
}
