% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDa__GetMapType}
\alias{p_GeoDa__GetMapType}
\title{p_GeoDa__GetMapType}
\usage{
p_GeoDa__GetMapType(xp)
}
\value{
The id of map type
}
\description{
Get the map type from GeoDa object
, which has to be one of {point_type, polygon_type, line_type, unkonwn_type}
}
