% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDa__GetNullValues}
\alias{p_GeoDa__GetNullValues}
\title{p_GeoDa__GetNullValues}
\usage{
p_GeoDa__GetNullValues(xp, col_name)
}
\value{
A list of string values
}
\description{
Get the null flags of a col from GeoDa object
}
