% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDa__GetNumCols}
\alias{p_GeoDa__GetNumCols}
\title{p_GeoDa__GetNumCols}
\usage{
p_GeoDa__GetNumCols(xp)
}
\value{
The number of columns
}
\description{
Get the number of columns from GeoDa object
}
