% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDa__GetNumericCol}
\alias{p_GeoDa__GetNumericCol}
\title{p_GeoDa__GetNumericCol}
\usage{
p_GeoDa__GetNumericCol(xp, col_name)
}
\value{
A list of double values
}
\description{
Get the values of a numeric col from GeoDa object
}
