% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDa__GetNumObs}
\alias{p_GeoDa__GetNumObs}
\title{p_GeoDa__GetNumObs}
\usage{
p_GeoDa__GetNumObs(xp)
}
\value{
The number of observations
}
\description{
Get the number of obervations from GeoDa object
}
