% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDa__GetStringCol}
\alias{p_GeoDa__GetStringCol}
\title{p_GeoDa__GetStringCol}
\usage{
p_GeoDa__GetStringCol(xp, col_name)
}
\value{
A list of string values
}
\description{
Get the values of a string col from GeoDa object
}
