% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDa__new}
\alias{p_GeoDa__new}
\title{p_GeoDa__new}
\usage{
p_GeoDa__new(file_path)
}
\arguments{
\item{file_path}{The path of the spatial file (ESRI shapefile)}
}
\value{
An externptr of GeoDa object
}
\description{
Create an instance of GeoDa by loading a ESRI shapefile
}
