% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDa__new1}
\alias{p_GeoDa__new1}
\title{p_GeoDa__new1}
\usage{
p_GeoDa__new1(layer_name, map_type, num_features, wkbs, wkb_bytes_len)
}
\arguments{
\item{file_path}{The path of the spatial file (ESRI shapefile)}
}
\value{
An externptr of GeoDa object
}
\description{
Create an instance of GeoDa from sf object
}
