% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoda.R
\docType{class}
\name{p_GeoDaTable-class}
\alias{p_GeoDaTable-class}
\alias{p_GeoDaTable}
\alias{$,p_GeoDaTable-method}
\title{p_GeoDaTable}
\description{
p_GeoDaTable class is a RefClass that wraps the C++ 'GeoDaTable' class.
See C++ functions in rcpp_rgeoda.cpp
}
