% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__GetDensity}
\alias{p_GeoDaWeight__GetDensity}
\title{p_GeoDaWeight__GetDensity}
\usage{
p_GeoDaWeight__GetDensity(xp)
}
\value{
The sparsity value
}
\description{
Get the density value of weights
}
