% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__GetMaxNeighbors}
\alias{p_GeoDaWeight__GetMaxNeighbors}
\title{p_GeoDaWeight__GetMaxNeighbors}
\usage{
p_GeoDaWeight__GetMaxNeighbors(xp)
}
\value{
The maximum number of neighbors
}
\description{
Get the maximum number of neighbors
}
