% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__GetMeanNeighbors}
\alias{p_GeoDaWeight__GetMeanNeighbors}
\title{p_GeoDaWeight__GetMeanNeighbors}
\usage{
p_GeoDaWeight__GetMeanNeighbors(xp)
}
\value{
The mean number of neighbors
}
\description{
Get the mean number of neighbors
}
