% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__GetMedianNeighbors}
\alias{p_GeoDaWeight__GetMedianNeighbors}
\title{p_GeoDaWeight__GetMedianNeighbors}
\usage{
p_GeoDaWeight__GetMedianNeighbors(xp)
}
\value{
The median number of neighbors
}
\description{
Get the median number of neighbors
}
