% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__GetMinNeighbors}
\alias{p_GeoDaWeight__GetMinNeighbors}
\title{p_GeoDaWeight__GetMinNeighbors}
\usage{
p_GeoDaWeight__GetMinNeighbors(xp)
}
\value{
The minimum number of neighbors
}
\description{
Get the minimum number of neighbors
}
