% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__GetNeighbors}
\alias{p_GeoDaWeight__GetNeighbors}
\title{p_GeoDaWeight__GetNeighbors()}
\usage{
p_GeoDaWeight__GetNeighbors(xp, obs_idx)
}
\value{
The list of neighbors
}
\description{
Get the neighbors of an observation
}
