% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__GetNeighborSize}
\alias{p_GeoDaWeight__GetNeighborSize}
\title{p_GeoDaWeight__GetNeighborSize()}
\usage{
p_GeoDaWeight__GetNeighborSize(xp, obs_idx)
}
\value{
The size of neighbors
}
\description{
Get the size of neighbors of an observation
}
