% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__GetNeighborWeights}
\alias{p_GeoDaWeight__GetNeighborWeights}
\title{p_GeoDaWeight__GetNeighborWeights()}
\usage{
p_GeoDaWeight__GetNeighborWeights(xp, obs_idx)
}
\value{
The list of weights values of neighbors
}
\description{
Get the weights value of neighbors of an observation
}
