% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__GetNumObs}
\alias{p_GeoDaWeight__GetNumObs}
\title{p_GeoDaWeight__GetNumObs}
\usage{
p_GeoDaWeight__GetNumObs(xp)
}
\value{
The number of observations
}
\description{
Get the number of observations from GeoDaWeight object
}
