% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__GetSparsity}
\alias{p_GeoDaWeight__GetSparsity}
\title{p_GeoDaWeight__GetSparsity}
\usage{
p_GeoDaWeight__GetSparsity(xp)
}
\value{
The sparsity value
}
\description{
Get the sparsity value of weights
}
