% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__HasIsolations}
\alias{p_GeoDaWeight__HasIsolations}
\title{p_GeoDaWeight__HasIsolations}
\usage{
p_GeoDaWeight__HasIsolations(xp)
}
\value{
Boolean if weights has any isolations
}
\description{
Return if the weights has any isolations
}
