% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__IsSymmetric}
\alias{p_GeoDaWeight__IsSymmetric}
\title{p_GeoDaWeight__IsSymmetric}
\usage{
p_GeoDaWeight__IsSymmetric(xp)
}
\value{
Boolean if weights is symmetric
}
\description{
Return if the weights is symmetric
}
