% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__SaveToFile}
\alias{p_GeoDaWeight__SaveToFile}
\title{p_GeoDaWeight__SaveToFile()}
\usage{
p_GeoDaWeight__SaveToFile(xp, out_path, layer_name, id_name, id_vec)
}
\value{
Boolean if successfully save the weights to a file
}
\description{
Save the weights to a file
}
