% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_GeoDaWeight__SpatialLag}
\alias{p_GeoDaWeight__SpatialLag}
\title{p_GeoDaWeight__GetSpatialLag}
\usage{
p_GeoDaWeight__SpatialLag(xp, obs_idx, vals)
}
\value{
The value of spatial lag of idx-th observation
}
\description{
Get the spatial lag value of idx-th observation
}
