% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoda.R
\name{p_GeoDaWeight_method}
\alias{p_GeoDaWeight_method}
\title{GeoDa_method, helper function to generate C functions
e.g. p_GeoDa__GetNumObs
Methods are listed in RcppExports.R}
\usage{
p_GeoDaWeight_method(name)
}
\description{
GeoDa_method, helper function to generate C functions
e.g. p_GeoDa__GetNumObs
Methods are listed in RcppExports.R
}
