% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoda.R
\docType{class}
\name{p_LISA-class}
\alias{p_LISA-class}
\alias{p_LISA}
\alias{$,p_LISA-method}
\title{p_LISA}
\description{
p_LISA class is a RefClass that wraps the C++ LISA class
See C++ functions in rcpp_lisa.cpp
}
