% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoda.R
\name{p_LISA_method}
\alias{p_LISA_method}
\title{LISA_method, helper function to generate C functions
e.g. p_LISA_Run
Methods are listed in RcppExports.R}
\usage{
p_LISA_method(name)
}
\description{
LISA_method, helper function to generate C functions
e.g. p_LISA_Run
Methods are listed in RcppExports.R
}
