% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_naturalbreaks}
\alias{p_naturalbreaks}
\title{p_naturalbreaks}
\usage{
p_naturalbreaks(k, data)
}
\arguments{
\item{k}{A numeric value indicates how many breaks}

\item{data}{A numeric vector}
}
\value{
A numeric vector
}
\description{
Natural Breaks group data whose boundaries are set where there are relatively big differences.
}
