% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{pca}
\alias{pca}
\title{Princinple Components Analysis}
\usage{
pca(data)
}
\arguments{
\item{data}{A 2d numeric list of selected variables}
}
\value{
A PCAResult object containing the details of PCA
}
\description{
Apply princinple components analysis on original data
}
