% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_geoda.R
\name{random_string}
\alias{random_string}
\title{random string}
\usage{
random_string(n = 5000)
}
\arguments{
\item{n}{a positive number, the number of items to choose from.}
}
\description{
Create a random string
}
