% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{read_gal}
\alias{read_gal}
\title{Read a .GAL file}
\usage{
read_gal(file_path, id_vec = c())
}
\arguments{
\item{file_path}{The file paht of the .GAL file}

\item{id_vec}{The id_vec is the id values used in the .GAL file. Default is empty.}
}
\value{
A weights object
}
\description{
Create a spatial weights object from a .GAL file
}
