% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{read_gwt}
\alias{read_gwt}
\title{Read a .GWT file}
\usage{
read_gwt(file_path, id_vec = c())
}
\arguments{
\item{file_path}{The file paht of the .GWT file}

\item{id_vec}{The id_vec is the id values used in the .GWT file.
Default is empty.}
}
\value{
A weights object
}
\description{
Create a spatial weights object from a .GWT file
}
