% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{read_swm}
\alias{read_swm}
\title{Read a .SWM file}
\usage{
read_swm(file_path, id_vec = numeric())
}
\arguments{
\item{file_path}{The file paht of the .SWM file}

\item{id_vec}{The id_vec is the id values used in the .SWM file. e.g. c(0,1,2,3,...)}
}
\value{
A weights object
}
\description{
Create a spatial weights object from a .SWM file
}
