% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{rook_weights}
\alias{rook_weights}
\title{Create a Rook contiguity weights}
\usage{
rook_weights(geoda_obj, ...)
}
\arguments{
\item{geoda_obj}{An object of [geoda] class}

\item{order}{(Optional) Order of contiguity}

\item{include_lower_order}{(Optional)  Whether or not the lower order neighbors should be included in the weights structure}

\item{precision_threshold}{(Optional) The precision of the underlying shape file is insufficient to allow for an exact match of coordinates to determine which polygons are neighbors}
}
\value{
w An object of Weight class
}
\description{
Create a Rook contiguity weights with options of "order", "include lower order" and "precision threshold"
}
