% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{set_neighbors}
\alias{set_neighbors}
\title{Set neighbors of an observation}
\usage{
set_neighbors(gda_w, idx, nbrs)
}
\arguments{
\item{gda_w}{A Weight object}

\item{idx}{A value indicates idx-th observation, idx start from 1}

\item{nbrs}{A list indicates the neighbors of idx-th observation (id start from 1)}
}
\description{
Set neighbors for idx-th observation, idx starts from 1
}
\examples{
\dontrun{
new_w <- create_weights(10)
set_neighbors(new_w, 1, c(2,3))
update_weights(new_w)
}
}
