% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize}
\alias{standardize}
\title{Standardize Data (Z)}
\usage{
standardize(data)
}
\arguments{
\item{data}{An input data for median absolute deviation}
}
\value{
A data array
}
\description{
Standarize data by transforming data to have zero mean and unit variance
}
