% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{summary.Weight}
\alias{summary.Weight}
\title{Summary of Spatial Weights}
\usage{
\method{summary}{Weight}(w_obj)
}
\arguments{
\item{gda_w}{A Weight object}
}
\description{
Override the summary() function for spatial weights
}
