% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{total_sumofsquare}
\alias{total_sumofsquare}
\title{Total Sum of Square}
\usage{
total_sumofsquare(data)
}
\arguments{
\item{data}{A list of numeric vectors which is used in spatial clustering methods, e.g. skater(k, w, data)}
}
\value{
A value of total sum of square value
}
\description{
Compute total sum of square value of a 2d tuple data
}
