% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_density}
\alias{weights_density}
\title{Density of a weights matrix}
\usage{
weights_density(gda_w)
}
\arguments{
\item{gda_w}{A Weight object}
}
\value{
Value of the weight matrix density
}
\description{
Get density computed from weights matrix
}
