% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_sparsity}
\alias{weights_sparsity}
\title{Sparsity of a weights matrix}
\usage{
weights_sparsity(gda_w)
}
\arguments{
\item{gda_w}{A Weight object}
}
\value{
Value of the weight matrix sparsity
}
\description{
Get sparsity computed from weights matrix
}
