% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geoda.R
\name{as.data.frame.geoda}
\alias{as.data.frame.geoda}
\title{convert rgeoda instance to data.frame}
\usage{
\method{as.data.frame}{geoda}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A rgeoda object}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{optional parameters}

\item{...}{other arguments passed to methods}
}
\value{
A data.frame object
}
\description{
Override the as.data.frame function for rgeoda instance
}
