% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demean}
\alias{demean}
\title{Demean standardize}
\usage{
demean(data)
}
\arguments{
\item{data}{An input data for median absolute deviation}
}
\value{
A data array
}
\description{
The mean for each variable is subtracting from each observation resulting in mean zero.
}
