% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geoda.R
\name{geoda_open}
\alias{geoda_open}
\title{Create a geoda object by reading a spatial dataset}
\usage{
geoda_open(ds_path)
}
\arguments{
\item{ds_path}{(character) The path of the spatial dataset}
}
\value{
gda_obj An object of geoda instance
}
\description{
Create a geoda object by reading a spatial dataset. The dataset that rgeoda supports includes: ESRI Shapefile, MapInfo File, CSV, GML, GPX, KML, GeoJSON, TopoJSON, OpenFileGDB, GFT Google Fusion Tables, CouchDB
}
