% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Get neighbors for idx-th observation based on weights matrix}
\usage{
get_neighbors(gda_w, idx)
}
\arguments{
\item{gda_w}{A Weight object}

\item{idx}{A value indicates idx-th observation, idx start from 1}
}
\value{
Vector of the neighbor indices, idx start from 1
}
\description{
Get neighbors for idx-th observation, idx starts from 1
}
