% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{has_isolates}
\alias{has_isolates}
\title{Check if weights matrix has isolates}
\usage{
has_isolates(gda_w)
}
\arguments{
\item{gda_w}{A Weight object}
}
\value{
Boolean value if weights matrix is symmetric
}
\description{
Check if weights matrix has isolates, or if any observation has no neighbors
}
