% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{is_symmetric}
\alias{is_symmetric}
\title{Check if weights matrix is symmetric}
\usage{
is_symmetric(gda_w)
}
\arguments{
\item{gda_w}{A Weight object}
}
\value{
Boolean value if weights matrix is symmetric
}
\description{
Check if weights matrix is symmetric
}
