% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{lisa_clusters}
\alias{lisa_clusters}
\title{Get values of local cluster indicators}
\usage{
lisa_clusters(gda_lisa, cutoff = 0.05)
}
\arguments{
\item{gda_lisa}{An instance of LISA object}

\item{cutoff}{A value of cutoff for significance p-values to filter not-significant clusters, default=0.05}
}
\value{
data a tuple values
}
\description{
Get the local cluster indicators returned from LISA computation.
}
