% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{lisa_fdr}
\alias{lisa_fdr}
\title{Get False Discovery Rate value of local spatial autocorrelation}
\usage{
lisa_fdr(gda_lisa, current_p)
}
\arguments{
\item{gda_lisa}{An instance of LISA object}

\item{current_p}{A value of current siginificant p-value}
}
\value{
A value of False Discovery Rate
}
\description{
Get False Discovery Rate value based on current LISA computation and current significat p-value
}
\examples{
\dontrun{
guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- geoda_open(guerry_path)
queen_w <- queen_weights(guerry)
guerry_df <- as.data.frame(guerry) # use as data.frame
crm_prs <- guerry_df['Crm_prs'][,1] # get values of variable "crm_prs"
lisa <- local_moran(queen_w, crm_prs)
fdr <- lisa_fdr(lisa, 0.05)
fdr
}
}
