% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{lisa_fdr}
\alias{lisa_fdr}
\title{Get False Discovery Rate value of local spatial autocorrelation}
\usage{
lisa_fdr(gda_lisa, current_p)
}
\arguments{
\item{gda_lisa}{An instance of LISA object}

\item{current_p}{A value of current siginificant p-value}
}
\value{
A value of False Discovery Rate
}
\description{
Get False Discovery Rate value based on current LISA computation and current significat p-value
}
