% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{lisa_labels}
\alias{lisa_labels}
\title{Get cluster labels}
\usage{
lisa_labels(gda_lisa)
}
\arguments{
\item{gda_lisa}{An instance of LISA object}
}
\value{
labels a tuple of values
}
\description{
Get cluster labels of LISA computation.
}
