% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{lisa_num_nbrs}
\alias{lisa_num_nbrs}
\title{Get numbers of neighbors of all observations}
\usage{
lisa_num_nbrs(gda_lisa)
}
\arguments{
\item{gda_lisa}{An instance of LISA object}
}
\value{
data a tuple of values
}
\description{
Get numbers of neighbors of all observations
}
