% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_multigeary}
\alias{local_multigeary}
\title{Local Multivariate Geary statistics}
\usage{
local_multigeary(
  w,
  data,
  permutations = 999,
  significance_cutoff = 0.05,
  cpu_threads = 6,
  seed = 123456789
)
}
\arguments{
\item{w}{An instance of Weight object}

\item{data}{A 2D tuple of values of selected variables}

\item{permutations}{The number of permutations for the LISA computation}

\item{significance_cutoff}{A cutoff value for significance p-values to filter not-significant clusters}

\item{cpu_threads}{The number of cpu threads used for parallel LISA computation}

\item{seed}{The seed for random number generator}
}
\value{
lisa_obj An instance of LISA (LocalSpatialAutocorrelation) object
}
\description{
The function to apply local Multivariate Geary statistics
}
