% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_multigeary}
\alias{local_multigeary}
\title{Local Multivariate Geary statistics}
\usage{
local_multigeary(w, data, ncpu = 8, perm = 999)
}
\arguments{
\item{w}{An instance of Weight object}

\item{data}{A data.frame with selected variables}

\item{ncpu}{The number of how many cpu threads being used, default: 8}

\item{perm}{The number of permutations used for permutation inference, default: 999}
}
\value{
lisa_obj An instance of LISA (LocalSpatialAutocorrelation) object
}
\description{
The function to apply local Multivariate Geary statistics
}
