% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisa.R
\name{local_quantilelisa}
\alias{local_quantilelisa}
\title{Quantile LISA Statistics}
\usage{
local_quantilelisa(
  w,
  df,
  k,
  q,
  permutations = 999,
  permutation_method = "complete",
  significance_cutoff = 0.05,
  cpu_threads = 6,
  seed = 123456789
)
}
\arguments{
\item{w}{An instance of Weight object}

\item{df}{A data frame with selected variable only. E.g. guerry["Crm_prs"]}

\item{k}{A value indicates the number of quantiles. Value range e.g. [1, 10]}

\item{q}{A value indicates which quantile or interval used in local join
count statistics. Value stars from 1.}

\item{permutations}{(optional) The number of permutations for the LISA
computation}

\item{permutation_method}{(optional) The permutation method used for the LISA
computation. Options are ('complete', 'lookup'). Default is 'complete'.}

\item{significance_cutoff}{(optional) A cutoff value for significance
p-values to filter not-significant clusters}

\item{cpu_threads}{(optional) The number of cpu threads used for parallel
LISA computation}

\item{seed}{(optional) The seed for random number generator}
}
\value{
An instance of LISA-class
}
\description{
The function to apply quantile LISA statistics
}
\examples{
library(sf)
guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- st_read(guerry_path)
queen_w <- queen_weights(guerry)
lisa <- local_quantilelisa(queen_w, guerry["Crm_prs"], k=4, q=1)
clsts <- lisa_clusters(lisa)
clsts
}
