% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{maxp}
\alias{maxp}
\title{An algorithm to solve the max-p-region problem}
\usage{
maxp(
  w,
  data,
  bound_vals,
  min_bound,
  local_search = "greedy",
  iterations = 99,
  tabu_length = 10,
  conv_tabu = 10,
  cool_rate = 0.85,
  sa_maxit = 1,
  initial_regions = c(0),
  distance_method = "euclidean",
  random_seed = 123456789
)
}
\arguments{
\item{w}{An instance of Weight class}

\item{data}{A 2d numeric list of selected variable}

\item{bound_vals}{A 1-d vector of selected bounding variable}

\item{min_bound}{A minimum value that the sum value of bounding variable int each cluster should be greater than}

\item{local_search}{The name of the heurist algorithm to find a optimal solution. Default to "greedy". Options are "greedy", "tabu" and "sa" (simulated annealing)}

\item{iterations}{(optional): The number of iterations of greedy algorithm. Defaults to 99.}

\item{tabu_length}{(optional): The length of a tabu search heuristic of tabu algorithm. Defaults to 10.}

\item{conv_tabu}{(optional): The number of non-improving moves. Defaults to 10.}

\item{cool_rate}{(optional): The cooling rate of a simulated annealing algorithm. Defaults to 0.85}

\item{sa_maxit}{(optional): The number of iterations of simulated annealing. Defaults to 1}

\item{distance_method}{(optional) The distance method used to compute the distance betwen observation i and j. Defaults to "euclidean". Options are "euclidean" and "manhattan"}

\item{random_seed}{(optional) The seed for random number generator. Defaults to 123456789.}

\item{init_regions}{(optional): The initial regions that the local search starts with. Default is empty. means the local search starts with a random process to "grow" clusters}
}
\value{
A 2d list represents a group of clusters
}
\description{
The max-p-region problem is a special case of constrained clustering where a finite number of geographical areas, n, are aggregated into the maximum number of regions, p, such that each region satisfies the following const raints: 1. The areas within a region must be geographically connected.
}
