% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{median_absolute_deviation}
\alias{median_absolute_deviation}
\title{Median absolute deviation}
\usage{
median_absolute_deviation(data)
}
\arguments{
\item{data}{An input data for median absolute deviation}
}
\value{
A data array
}
\description{
Median absolute deviation to measure  measure of the variability of a univariate sample of quantitative data
}
