% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{natural_breaks}
\alias{natural_breaks}
\title{Natural Breaks (Jenks)}
\usage{
natural_breaks(k, data)
}
\arguments{
\item{k}{A numeric value indicates how many breaks}

\item{data}{An input data}
}
\value{
A array contains values of computed breaks
}
\description{
Natural Breaks group data whose boundaries are set where there are relatively big differences.
}
