% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeoda.R
\docType{class}
\name{p_GeoDa-class}
\alias{p_GeoDa-class}
\alias{p_GeoDa}
\alias{$,p_GeoDa-method}
\title{p_GeoDa}
\description{
p_GeoDa class is a RefClass that wraps the C++ GeoDa class
See C++ functions in rcpp_rgeoda.cpp
}
