% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percentile_breaks}
\alias{percentile_breaks}
\title{Percentile Breaks}
\usage{
percentile_breaks(k, data)
}
\arguments{
\item{k}{A numeric value indicates how many breaks}

\item{data}{An input data}
}
\value{
A array contains values of computed breaks
}
\description{
Percentile breaks data into 6 groups: the lowest 1%, 1-10%, 10-50%, 50-90%, 90-99% and the top 1%.
}
