% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_geoda.R
\name{sf_to_geoda}
\alias{sf_to_geoda}
\title{Create an instance of geoda-class from a 'sf' object}
\usage{
sf_to_geoda(sf_obj, with_table = TRUE)
}
\arguments{
\item{sf_obj}{An instance of 'sf' object}

\item{with_table}{A boolean flag indicates if table is copied from sf object
to create geoda object. Default is TRUE.}
}
\value{
An instance of geoda-class
}
\description{
Create an instance of geoda-class from a 'sf' object returned
from 'st_read()' function. NOTE: The table content is NOT used to create an
instance of geoda-class.
}
