% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_geoda.R
\name{sp_to_geoda}
\alias{sp_to_geoda}
\title{Create a geoda object from a sp object}
\usage{
sp_to_geoda(sp_obj, with_table = FALSE)
}
\arguments{
\item{sp_obj}{An instance of sp object}

\item{with_table}{Optional, Default: FALSE If create a table from sp dataframe object.}
}
\value{
geoda_obj An instance of GeoDa object
}
\description{
The sp package has been an essential tool which provides spatial data-structures and many utility functions to do spatial analysis in R. It has been a core dependent library for many other packages, e.g. rgdal (IO), maptools (mapping), spdep (spatial weights, spatial statistics, and spatial models) etc.
Using rgdal to read a ESRI Shapefile will return a sp (Spatial object) object, which could be either a SpatialPointsDataFrame (using an AttributeList for its data slot directly), a SpatialLinesDataFrame, or a SpatialPolygonsDataFrame.
}
