% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stddev_breaks}
\alias{stddev_breaks}
\title{Standard Deviation Breaks}
\usage{
stddev_breaks(k, data)
}
\arguments{
\item{k}{A numeric value indicates how many breaks}

\item{data}{An input data}
}
\value{
A array contains values of computed breaks
}
\description{
Standard deviation breaks first transforms data to standard deviation units (mean=0, stddev=1), and then divide the range of values into 6 groups.
}
