% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{summary._p_PCAResult}
\alias{summary._p_PCAResult}
\title{Summary of Princinple Components Analysis Result}
\usage{
\method{summary}{`_p_PCAResult`}(object, ...)
}
\arguments{
\item{object}{A PCAResult object}

\item{...}{summary optional parameters}
}
\value{
A summary  description of PCA object
}
\description{
Override the summary() function for a PCA result
}
