% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{summary.Weight}
\alias{summary.Weight}
\title{Summary of Spatial Weights}
\usage{
\method{summary}{Weight}(object, ...)
}
\arguments{
\item{object}{A Weight object}

\item{...}{summary optional parameters}
}
\value{
A summary description of Weight object
}
\description{
Override the summary() function for spatial weights
}
\examples{
\dontrun{
guerry_path <- system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry <- geoda_open(guerry_path)
queen_w <- queen_weights(guerry)
summary(queen_w)
}
}
